+++ src/components/todayTasksComponent/TodayTasksComponent.js
@@ -1,20 +1,45 @@
 import React, { useState } from 'react'
-import todayTaskStyle from './TodayTasksComponent.module.scss';
-import OpenedTasks from '../openedTaskComponent/OpenedTasksComponent';
+import todayTaskStyle from '../todayTasksComponent/TodayTasksComponent.module.scss';
 
 const TodayTask = (props) => {
-    const tasksForRender = props.dayForRender.arrayOfTasks;
-    const [openedStatusClassName, setOpenedStatusClassName] = useState('closedStatusClassName');
-    const [selectedDayForRender, setSelectedDayForRender] = useState(tasksForRender);
+    const [openedStatusClassName, setOpenedStatusClassName] = useState('openedStatusClassName');
+    const [selectedDayForRender, setSelectedDayForRender] = useState(props.dayForRender);
 
-    function handlerButtonClick(e) {
-        const card = e.target.parentNode.nextSibling;
-        if (openedStatusClassName == 'closedStatusClassName') {
-            card.style.display = 'flex'
-            setOpenedStatusClassName('openedStatusClassName');
+    const openedTasks = selectedDayForRender.map( (task) =>
+        <div className={todayTaskStyle.tasksCard} key={task.color + task.name + 222}>
+            <div className={task.color} key={task.name + 333 + task.color} />
+            <div 
+                className={todayTaskStyle.textOfOpenedTask}
+                key={task.name + 222 + task.color} 
+            >
+                <h3 key={task.description + task.name + 'j' + 0}>
+                    {task.name}
+                </h3>
+                <p key={task.name + 'j' + task.description + 3}>
+                    {task.description}
+                </p>
+            </div>
+                <label className='switch'>
+                    <input type="checkbox" className='switchInput'/>
+                    <span className='slider round'  onClick={handleSwitchClick}></span>
+                </label>
+        </div>
+    )
+
+    function handleSwitchClick(e) {
+        const checkedText = e.target.parentNode.parentNode.children[1];
+        if (checkedText.style.textDecoration == 'line-through') {
+            checkedText.style.textDecoration = 'none'
         } else {
-            setOpenedStatusClassName('closedStatusClassName');
-            card.style.display = 'none'
+            checkedText.style.textDecoration = 'line-through'
+        }
+    }
+
+    function handlerButtonClick() {
+        if (openedStatusClassName == 'closedStatusClassName') {
+            setOpenedStatusClassName('openedStatusClassName');    
+        } else {
+            setOpenedStatusClassName('closedStatusClassName'); 
         }
     }
 
@@ -25,9 +50,9 @@ const TodayTask = (props) => {
                     className={openedStatusClassName} 
                     onClick={handlerButtonClick} 
                 />
-                <h3>{props.dayForRender.title + " "}:</h3>
+                <h3>Today Task:</h3>
             </div>
-            <OpenedTasks openedTasks = {selectedDayForRender}/>
+            <div className={todayTaskStyle.openedTasks}>{openedTasks}</div>
         </div>
     )
 }
