+++ src/root/App.scss
@@ -30,6 +30,10 @@ body{
         font-size: 35px;
         color: #fff;
     }
+
+    img{
+        cursor: pointer;
+    }
 }
 
 .redLine{
